Reorganizing fix library
