+++ fix/fixwin32.h
