+++ vecmat/CMakeLists.txt
@@ -3,3 +3,4 @@ set(CPPS
   vector.cpp)
 
 add_library(vecmat STATIC ${HEADERS} ${CPPS})
+target_link_libraries(vecmat fix)
